Add a Keymap class
The job of the Keymap class is to keep track of the keybinding map, that is,
a map from a key to an associated action. The Keymap class can then take a
key event, find the associated action, and invoke it. The code is more
verbose than usual due to the need to map from X's key semantics to something
less elegantly orthogonal but more useful to application developers.