Add code to load config files
By default, an RTK program will attempt to load a config file from /etc/rtk
and from the user's home directory. If the latter is not found, it is created
for the benefit of the user. RTK comes with a built-in default config file,
built with yaml2c.pl, which is used to create the file in the user's home
directory, and is also used to fill in any values that are missing in the
user's file.